Divides ConfigForm to multiple components
